# services/go/payment/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  description: |
    NOWPayments integration API for crypto deposits and withdrawals on a gambling platform.
    
    ## Authentication
    All endpoints require Bearer token authentication via the `Authorization` header.
    
    ## Rate Limiting
    - 100 requests per minute per user
    - 1000 requests per minute per IP for webhook endpoints
    
    ## Idempotency
    All write operations support idempotency via the `idempotency_key` field.
    Replaying a request with the same key returns the original response.
    
    ## Error Codes
    Payment service errors use codes in the 5000-5999 range.
  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@example.com
servers:
  - url: https://api.example.com
    description: Production
  - url: https://api-staging.example.com
    description: Staging
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/payments/withdraw:
    post:
      tags:
        - Withdrawals
      summary: Initiate a crypto withdrawal
      description: |
        Creates a withdrawal request to send crypto to an external address.
        
        Requirements:
        - KYC level 2 or higher required
        - Sufficient available balance
        - Valid crypto address for the selected currency
        
        Daily withdrawal limits by KYC level:
        - Level 0: $0/day (not allowed)
        - Level 1: $500/day
        - Level 2: $5,000/day
        - Level 3: $25,000/day
      operationId: initiateWithdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
            examples:
              btc_withdrawal:
                summary: BTC Withdrawal
                value:
                  amount: "50.00"
                  currency: "BTC"
                  address: "bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh"
                  idempotency_key: "wdh_abc123xyz789"
              usdt_withdrawal:
                summary: USDT TRC-20 Withdrawal
                value:
                  amount: "100.00"
                  currency: "USDTTRX"
                  address: "TRX1234567890abcdefghijklmnopqrstuvwxyz"
                  idempotency_key: "wdh_def456uvw012"
      responses:
        '201':
          description: Withdrawal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawResponse'
              example:
                success: true
                data:
                  withdrawal_uuid: "770g0600-g41d-63f6-c938-668877662222"
                  withdrawal_id: "987654321"
                  amount: "0.00125"
                  fiat_amount: "50.00"
                  currency: "BTC"
                  address: "bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh"
                  status: "processing"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_address:
                  summary: Invalid crypto address
                  value:
                    code: 5008
                    message: "invalid crypto address format"
                invalid_amount:
                  summary: Invalid amount
                  value:
                    code: 5014
                    message: "invalid amount"
        '403':
          description: KYC level insufficient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 5004
                message: "KYC level 2 required for withdrawals"
                details:
                  current_level: 1
                  required_level: 2
        '422':
          description: Insufficient balance or limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficient_balance:
                  summary: Insufficient balance
                  value:
                    code: 5006
                    message: "insufficient balance"
                    details:
                      available: 25.00
                      requested: 50.00
                limit_exceeded:
                  summary: Daily limit exceeded
                  value:
                    code: 5005
                    message: "daily limit exceeded"
                    details:
                      limit: 500.00
                      used: 480.00
                      requested: 50.00
                      available: 20.00
  /api/v1/payments/withdrawals/{id}:
    get:
      tags:
        - Withdrawals
      summary: Get withdrawal by ID
      description: Retrieves a specific withdrawal by its UUID
      operationId: getWithdrawal
      parameters:
        - name: id
          in: path
          required: true
          description: Withdrawal UUID
          schema:
            type: string
            format: uuid
          example: "770g0600-g41d-63f6-c938-668877662222"
      responses:
        '200':
          description: Withdrawal found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'
              example:
                success: true
                data:
                  uuid: "770g0600-g41d-63f6-c938-668877662222"
                  withdrawal_id: "987654321"
                  user_id: 1001
                  amount: "0.00125"
                  fiat_amount: "50.00"
                  fiat_currency: "USD"
                  crypto_currency: "BTC"
                  address: "bc1qxy2k...0wlh"
                  status: "finished"
                  created_at: "2024-01-15T10:00:00Z"
                  updated_at: "2024-01-15T10:30:00Z"
                  completed_at: "2024-01-15T10:30:00Z"
        '404':
          description: Withdrawal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 5012
                message: "withdrawal not found"
                details:
                  withdrawal_id: "770g0600-g41d-63f6-c938-668877662222"
  /api/v1/payments/withdrawals/history:
    get:
      tags:
        - Withdrawals
      summary: List withdrawal history
      description: Retrieves paginated withdrawal history for the authenticated user
      operationId: getWithdrawalHistory
      parameters:
        - name: limit
          in: query
          description: Number of items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
        - name: cursor
          in: query
          description: Pagination cursor for next page
          schema:
            type: string
          example: "eyJpZCI6MTIzNDU2fQ=="
        - name: status
          in: query
          description: Filter by withdrawal status
          schema:
            $ref: '#/components/schemas/WithdrawalStatus'
      responses:
        '200':
          description: Withdrawal history retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalHistoryResponse'
              example:
                success: true
                data:
                  items:
                    - withdrawal_uuid: "770g0600-g41d-63f6-c938-668877662222"
                      amount: "50.00"
                      currency: "BTC"
                      address: "bc1qxy2k...0wlh"
                      status: "finished"
                      created_at: "2024-01-15T10:00:00Z"
                    - withdrawal_uuid: "880h1700-h52e-74g7-d049-779988773333"
                      amount: "100.00"
                      currency: "ETH"
                      address: "0x1234...5678"
                      status: "processing"
                      created_at: "2024-01-14T15:00:00Z"
                  next_cursor: "eyJpZCI6MTEwMDAwfQ=="
                  has_more: false
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication service
  schemas:
...
    WithdrawRequest:
      type: object
      required:
        - amount
        - currency
        - address
        - idempotency_key
      properties:
        amount:
          type: string
          description: Fiat amount in USD to withdraw
          pattern: '^\d+(\.\d{1,2})?$'
          example: "50.00"
        currency:
          type: string
          description: Cryptocurrency code
          enum:
            - BTC
            - ETH
            - USDTETH
            - USDTTRX
            - USDC
            - LTC
            - BCH
          example: "BTC"
        address:
          type: string
          description: Destination crypto address
          minLength: 26
          maxLength: 100
          example: "bc1qxy2kgdygjrsqtzq2n0yrf2493p83kkfjhx0wlh"
        idempotency_key:
          type: string
          description: Unique key for idempotent requests
          minLength: 1
          maxLength: 255
          example: "wdh_abc123xyz789"

    # Response schemas
...
    WithdrawResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            withdrawal_uuid:
              type: string
              format: uuid
              description: Internal withdrawal UUID
            withdrawal_id:
              type: string
              description: NOWPayments withdrawal ID
            amount:
              type: string
              description: Crypto amount
            fiat_amount:
              type: string
              description: Fiat amount in USD
            currency:
              type: string
              description: Cryptocurrency code
            address:
              type: string
              description: Destination address (may be masked)
            status:
              $ref: '#/components/schemas/WithdrawalStatus'
    WithdrawalResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Withdrawal'
    Withdrawal:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        withdrawal_id:
          type: string
        user_id:
          type: integer
          format: int64
        amount:
          type: string
        fiat_amount:
          type: string
        fiat_currency:
          type: string
        crypto_currency:
          type: string
        address:
          type: string
          description: Destination address (may be masked)
        status:
          $ref: '#/components/schemas/WithdrawalStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    WithdrawalHistoryResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/WithdrawalItem'
            next_cursor:
              type: string
              nullable: true
            has_more:
              type: boolean
    WithdrawalItem:
      type: object
      properties:
        withdrawal_uuid:
          type: string
          format: uuid
        amount:
          type: string
        currency:
          type: string
        address:
          type: string
          description: Destination address (masked)
        status:
          $ref: '#/components/schemas/WithdrawalStatus'
        created_at:
          type: string
          format: date-time

    # Enums
...
    WithdrawalStatus:
      type: string
      enum:
        - processing
        - sending
        - sent
        - finished
        - failed
        - cancelled
      description: |
        Withdrawal status values:
        - `processing`: Withdrawal request being processed
        - `sending`: Sending to destination address
        - `sent`: Transaction sent to blockchain
        - `finished`: Withdrawal completed successfully
        - `failed`: Withdrawal failed
        - `cancelled`: Withdrawal was cancelled

    # Webhook schemas
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Error code
          example: 5001
        message:
          type: string
          description: Human-readable error message
          example: "payment not found"
        details:
          type: object
          description: Additional error details
          additionalProperties: true

    # Error codes reference
...